page crash on cookie learn more Url
Add NULL/empty check for retrieval of the learnMoreUrl on the cookie policy popin. If the specified page has no result for specific scope (storeview), this method threw an exception. Changed to return "#" but perhaps it would be better to return NULL and let the caller of getLearnMoreUrl() decide what to do, default or hide the link